hotfix: refactor imports for hoc, add hoc to '/'
