+++ src/pages/auth/create-new-password/index.tsx
@@ -11,7 +11,7 @@ import {
 } from '@/feature/auth'
 import { AuthRoutes } from '@/shared/const/routes'
 import { handleErrorResponse } from '@/shared/helpers/handleErrorResponse'
-import { DefenderAuthRoute } from '@/shared/helpers/hoc/DefenderAuthRoute'
+import { DefenderAuthRoute } from '@/shared/helpers/hoc'
 import { UseFormRef } from '@/shared/types/form'
 import { Page } from '@/shared/types/layout'
 import { Loader } from '@/shared/ui/Loader'
