+++ src/pages/auth/forgot-password/index.tsx
@@ -8,7 +8,7 @@ import {
   useResendRecoveryPasswordMutation,
 } from '@/feature/auth'
 import { handleErrorResponse } from '@/shared/helpers/handleErrorResponse'
-import { DefenderAuthRoute } from '@/shared/helpers/hoc/DefenderAuthRoute'
+import { DefenderAuthRoute } from '@/shared/helpers/hoc'
 import { UseFormRef } from '@/shared/types/form'
 import { Page } from '@/shared/types/layout'
 import { AuthLayout } from '@/widgets/layout'
