+++ src/pages/auth/sign-in/index.tsx
@@ -3,7 +3,7 @@ import { useRef } from 'react'
 import { SignInForm, SignInFormValues, useLoginMutation } from '@/feature/auth'
 import { AppRoutes, GeneralRoutes } from '@/shared/const/routes'
 import { handleErrorResponse } from '@/shared/helpers/handleErrorResponse'
-import { DefenderAuthRoute } from '@/shared/helpers/hoc/DefenderAuthRoute'
+import { DefenderAuthRoute } from '@/shared/helpers/hoc'
 import { UseFormRef } from '@/shared/types/form'
 import { Page } from '@/shared/types/layout'
 import { AuthLayout } from '@/widgets/layout'
