+++ src/pages/messenger/index.tsx
@@ -1,6 +1,6 @@
 import { ReactElement } from 'react'
 
-import { DefenderProtectedRoute } from '@/shared/helpers/hoc/DefenderProtectedRoute'
+import { DefenderProtectedRoute } from '@/shared/helpers/hoc'
 import { Page } from '@/shared/types/layout'
 import { SidebarLayout } from '@/widgets/layout'
