+++ src/pages/profile/index.tsx
@@ -5,8 +5,7 @@ import { PostItem, PostsList, PostsListSkeleton } from '@/entities/post'
 import { ProfileInfo } from '@/entities/profile'
 import { PostDetailsDialogs, invalidateTagsPost, useGetMyPostsQuery } from '@/feature/post'
 import { useGetMyProfileQuery } from '@/feature/profile'
-import { DefenderAuthRoute } from '@/shared/helpers/hoc/DefenderAuthRoute'
-import { DefenderProtectedRoute } from '@/shared/helpers/hoc/DefenderProtectedRoute'
+import { DefenderProtectedRoute } from '@/shared/helpers/hoc'
 import { useInfinityScroll } from '@/shared/hooks/useInfinityScroll'
 import { Page } from '@/shared/types/layout'
 import { Loader } from '@/shared/ui/Loader'
