+++ src/shared/helpers/hoc/index.ts
@@ -1,0 +1,3 @@
+export { DefenderAuthRoute } from './DefenderAuthRoute'
+export { DefenderProtectedRoute } from './DefenderProtectedRoute'
+export { DefenderPublicRoute } from './DefenderPublicRoute'
